Move Compatibility Enums to own package
