+++ src/main/java/metadev/digital/metacustomitemslib/compatibility/addons/MobHuntingCompat.java
@@ -1,7 +1,7 @@
 package metadev.digital.metacustomitemslib.compatibility.addons;
 
 import metadev.digital.metacustomitemslib.Core;
-import metadev.digital.metacustomitemslib.compatibility.SupportedPluginEntities;
+import metadev.digital.metacustomitemslib.compatibility.enums.SupportedPluginEntities;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
