+++ src/main/java/metadev/digital/metacustomitemslib/compatibility/enums/BoundIdentifierEnum.java
@@ -1,4 +1,4 @@
-package metadev.digital.metacustomitemslib.compatibility;
+package metadev.digital.metacustomitemslib.compatibility.enums;
 
 /***
  * Enum identifier for whether a feature with a single presented SemVer should be treated as an upper or lower bound.
