+++ src/main/java/metadev/digital/metacustomitemslib/compatibility/enums/VersionSetIdentifierEnum.java
@@ -1,4 +1,4 @@
-package metadev.digital.metacustomitemslib.compatibility;
+package metadev.digital.metacustomitemslib.compatibility.enums;
 
 /***
  * Enum identifier for whether a feature with a single presented version set should be treated as a server or plugin version binding.
